--- a/cabal.project
+++ b/cabal.project
@@ -18,8 +18,8 @@ source-repository-package
 source-repository-package
     type: git
     location: https://github.com/kadena-io/chainweb-api.git
-    tag: c877d32b46175917d9df9521d0c4e7ec47ebd5cd
-    --sha256: sha256-X6fQTcV46lHT3vifSDO0exnoYBr7Tx2YEpK8EyEFgBA=
+    tag: 6725d8490fcf43fd5f6b8b0ca731565a58f9c8e6
+    --sha256: sha256-B3swFmWuEajZ42gpdvHsF2MpWxHU4YiIx/nV+1ZCx30=
 
 source-repository-package
     type: git