--- a/haskell-src/chainweb-data.cabal
+++ b/haskell-src/chainweb-data.cabal
@@ -87,6 +87,7 @@ library
     , optparse-applicative     >=0.14 && <0.17
     , servant-client
     , servant-openapi3
+    , vector
     , yet-another-logger
 
   if flag(ghc-flags)