--- a/haskell-src/exec/Chainweb/Server.hs
+++ b/haskell-src/exec/Chainweb/Server.hs
@@ -405,26 +405,8 @@ throw404 msg = throwError $ err404 { errBody = msg }
 throw400 :: MonadError ServerError m => ByteString -> m a
 throw400 msg = throwError $ err400 { errBody = msg }
 
-txHandler
-  :: LogFunctionIO Text
-  -> M.Managed Connection
-  -> Maybe RequestKey
-  -> Handler TxDetail
-txHandler _ _ Nothing = throw404 "You must specify a search string"
-txHandler logger pool (Just (RequestKey rk)) =
-  may404 $ liftIO $ M.with pool $ \c ->
-  runBeamPostgresDebug (logger Debug . T.pack) c $ do
-    r <- runSelectReturningOne $ select $ do
-      tx <- all_ (_cddb_transactions database)
-      blk <- all_ (_cddb_blocks database)
-      guard_ (_tx_block tx `references_` blk)
-      guard_ (_tx_requestKey tx ==. val_ (DbHash rk))
-      return (tx,blk)
-    evs <- runSelectReturningList $ select $ do
-       ev <- all_ (_cddb_events database)
-       guard_ (_ev_requestkey ev ==. val_ (RKCB_RequestKey $ DbHash rk))
-       return ev
-    return $ (`fmap` r) $ \(tx,blk) -> TxDetail
+toApiTxDetail :: Transaction -> Block -> [Event] -> TxDetail
+toApiTxDetail tx blk evs = TxDetail
         { _txDetail_ttl = fromIntegral $ _tx_ttl tx
         , _txDetail_gasLimit = fromIntegral $ _tx_gasLimit tx
         , _txDetail_gasPrice = _tx_gasPrice tx
@@ -450,15 +432,35 @@ txHandler logger pool (Just (RequestKey rk)) =
         , _txDetail_requestKey = unDbHash $ _tx_requestKey tx
         , _txDetail_sender = _tx_sender tx
         , _txDetail_code = _tx_code tx
-        , _txDetail_success =
-          maybe False (const True) $ _tx_goodResult tx
+        , _txDetail_success = isJust $ _tx_goodResult tx
         , _txDetail_events = map toTxEvent evs
         }
-
   where
     unMaybeValue = maybe Null unPgJsonb
     toTxEvent ev =
       TxEvent (_ev_qualName ev) (unPgJsonb $ _ev_params ev)
+
+txHandler
+  :: LogFunctionIO Text
+  -> M.Managed Connection
+  -> Maybe RequestKey
+  -> Handler TxDetail
+txHandler _ _ Nothing = throw404 "You must specify a search string"
+txHandler logger pool (Just (RequestKey rk)) =
+  may404 $ liftIO $ M.with pool $ \c ->
+  runBeamPostgresDebug (logger Debug . T.pack) c $ do
+    r <- runSelectReturningOne $ select $ do
+      tx <- all_ (_cddb_transactions database)
+      blk <- all_ (_cddb_blocks database)
+      guard_ (_tx_block tx `references_` blk)
+      guard_ (_tx_requestKey tx ==. val_ (DbHash rk))
+      return (tx,blk)
+    evs <- runSelectReturningList $ select $ do
+       ev <- all_ (_cddb_events database)
+       guard_ (_ev_requestkey ev ==. val_ (RKCB_RequestKey $ DbHash rk))
+       return ev
+    return $ (`fmap` r) $ \(tx,blk) -> toApiTxDetail tx blk evs
+  where
     may404 a = a >>= maybe (throw404 "Tx not found") return
 
 txsHandler
@@ -480,44 +482,11 @@ txsHandler logger pool (Just (RequestKey rk)) =
        ev <- all_ (_cddb_events database)
        guard_ (_ev_requestkey ev ==. val_ (RKCB_RequestKey $ DbHash rk))
        return ev
-    return $ (`fmap` r) $ \(tx,blk) -> TxDetail
-        { _txDetail_ttl = fromIntegral $ _tx_ttl tx
-        , _txDetail_gasLimit = fromIntegral $ _tx_gasLimit tx
-        , _txDetail_gasPrice = _tx_gasPrice tx
-        , _txDetail_nonce = _tx_nonce tx
-        , _txDetail_pactId = unDbHash <$> _tx_pactId tx
-        , _txDetail_rollback = _tx_rollback tx
-        , _txDetail_step = fromIntegral <$> _tx_step tx
-        , _txDetail_data = unMaybeValue $ _tx_data tx
-        , _txDetail_proof = _tx_proof tx
-        , _txDetail_gas = fromIntegral $ _tx_gas tx
-        , _txDetail_result =
-          maybe (unMaybeValue $ _tx_badResult tx) unPgJsonb $
-          _tx_goodResult tx
-        , _txDetail_logs = fromMaybe "" $ _tx_logs tx
-        , _txDetail_metadata = unMaybeValue $ _tx_metadata tx
-        , _txDetail_continuation = unPgJsonb <$> _tx_continuation tx
-        , _txDetail_txid = maybe 0 fromIntegral $ _tx_txid tx
-        , _txDetail_chain = fromIntegral $ _tx_chainId tx
-        , _txDetail_height = fromIntegral $ _block_height blk
-        , _txDetail_blockTime = _block_creationTime blk
-        , _txDetail_blockHash = unDbHash $ unBlockId $ _tx_block tx
-        , _txDetail_creationTime = _tx_creationTime tx
-        , _txDetail_requestKey = unDbHash $ _tx_requestKey tx
-        , _txDetail_sender = _tx_sender tx
-        , _txDetail_code = _tx_code tx
-        , _txDetail_success =
-          maybe False (const True) $ _tx_goodResult tx
-        , _txDetail_events = map toTxEvent evs
-        }
-
+    return $ (`fmap` r) $ \(tx,blk) -> toApiTxDetail tx blk evs
   where
     emptyList404 xs = xs >>= \case
       [] -> throw404 "no txs not found"
       ys ->  return ys
-    unMaybeValue = maybe Null unPgJsonb
-    toTxEvent ev =
-      TxEvent (_ev_qualName ev) (unPgJsonb $ _ev_params ev)
 
 type AccountNextToken = (Int64, T.Text, Int64)