--- a/haskell-src/exec/Chainweb/Server.hs
+++ b/haskell-src/exec/Chainweb/Server.hs
@@ -440,37 +440,8 @@ toApiTxDetail tx blk evs = TxDetail
     toTxEvent ev =
       TxEvent (_ev_qualName ev) (unPgJsonb $ _ev_params ev)
 
-txHandler
-  :: LogFunctionIO Text
-  -> M.Managed Connection
-  -> Maybe RequestKey
-  -> Handler TxDetail
-txHandler _ _ Nothing = throw404 "You must specify a search string"
-txHandler logger pool (Just (RequestKey rk)) =
-  may404 $ liftIO $ M.with pool $ \c ->
-  runBeamPostgresDebug (logger Debug . T.pack) c $ do
-    r <- runSelectReturningOne $ select $ do
-      tx <- all_ (_cddb_transactions database)
-      blk <- all_ (_cddb_blocks database)
-      guard_ (_tx_block tx `references_` blk)
-      guard_ (_tx_requestKey tx ==. val_ (DbHash rk))
-      return (tx,blk)
-    evs <- runSelectReturningList $ select $ do
-       ev <- all_ (_cddb_events database)
-       guard_ (_ev_requestkey ev ==. val_ (RKCB_RequestKey $ DbHash rk))
-       return ev
-    return $ (`fmap` r) $ \(tx,blk) -> toApiTxDetail tx blk evs
-  where
-    may404 a = a >>= maybe (throw404 "Tx not found") return
-
-txsHandler
-  :: LogFunctionIO Text
-  -> M.Managed Connection
-  -> Maybe RequestKey
-  -> Handler [TxDetail]
-txsHandler _ _ Nothing = throw404 "You must specify a search string"
-txsHandler logger pool (Just (RequestKey rk)) =
-  emptyList404 $ liftIO $ M.with pool $ \c ->
+queryTxsByKey :: LogFunctionIO Text -> Text -> Connection -> IO [TxDetail]
+queryTxsByKey logger rk c =
   runBeamPostgresDebug (logger Debug . T.pack) c $ do
     r <- runSelectReturningList $ select $ do
       tx <- all_ (_cddb_transactions database)
@@ -483,10 +454,28 @@ txsHandler logger pool (Just (RequestKey rk)) =
        guard_ (_ev_requestkey ev ==. val_ (RKCB_RequestKey $ DbHash rk))
        return ev
     return $ (`fmap` r) $ \(tx,blk) -> toApiTxDetail tx blk evs
-  where
-    emptyList404 xs = xs >>= \case
-      [] -> throw404 "no txs not found"
-      ys ->  return ys
+
+txHandler
+  :: LogFunctionIO Text
+  -> M.Managed Connection
+  -> Maybe RequestKey
+  -> Handler TxDetail
+txHandler _ _ Nothing = throw404 "You must specify a search string"
+txHandler logger pool (Just (RequestKey rk)) =
+  liftIO (M.with pool $ queryTxsByKey logger rk) >>= \case
+    [x] -> return x
+    _ -> throw404 "Tx not found"
+
+txsHandler
+  :: LogFunctionIO Text
+  -> M.Managed Connection
+  -> Maybe RequestKey
+  -> Handler [TxDetail]
+txsHandler _ _ Nothing = throw404 "You must specify a search string"
+txsHandler logger pool (Just (RequestKey rk)) =
+  liftIO (M.with pool $ queryTxsByKey logger rk) >>= \case
+    [] -> throw404 "Tx not found"
+    xs -> return xs
 
 type AccountNextToken = (Int64, T.Text, Int64)