This PR extends the transaction endpoints to gather the history of continuation transactions. It extends the following endpoints and adds the `initialCode` and `previousSteps` fields to them:
* Single transaction details: `/txs/tx`
* Multiple transaction details: `/txs/txs`
* Transaction search: `/txs/search`
* Recent transactions: `/txs/recent`

In addition to attaching this continuation history to the response, this PR also extends the search logic of `/txs/search` so that if the given search term appears in the `initialCode` of a continuation transaction, the search still finds it.

This PR supports https://github.com/kadena-io/block-explorer/issues/64 and https://github.com/kadena-io/block-explorer/issues/63