@enobayram 2023-02-27
@sirlensalot thanks a lot for this PR. I've just merged master into this branch and adapted your changes to work with the new bounded-scans. This way of searching involves an index lookup through `pactId = requestKey`, which is a random string, meaning all lookups will cause page faults, so I was worried about the performance implications and ran a comparison with `master`:

#### Search random string on master

Searching for a random string means we let the database scan through 50000 transactions before giving up with a continuation token.

```
$ time curl -v "$CWDURL/txs/search?maxheight=2900000&search=somerandomstring" | jq '.[] | { requestKey, height, code, pactId: .continuation.pactId }'

real	0m1,238s
```

#### Search random string on this branch

```
$ time curl -v "$CWDURL/txs/search?maxheight=2800000&search=somerandomstring" | jq '.[] | { requestKey, height, code, pactId: .continuation.pactId }'

real	0m1,407s
```

So it seems like this branch is only slightly slower. When I run this test with different `maxheight`s (since otherwise all disk reads would be cached), the timing jumps between 0.8s to 1.5s, but it seems to me like this branch is consistently just a little slower, but I don't think the difference matters too much.

### Testing the feature

I've manually tested to see whether this branch is able to find [transaction 6nN3D_yO8uwDuoyRnD44eeEGOz8W5UCBVr2GYRrVSKI](https://explorer.chainweb.com/mainnet/txdetail/6nN3D_yO8uwDuoyRnD44eeEGOz8W5UCBVr2GYRrVSKI) which is a continuation of [J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ](https://explorer.chainweb.com/mainnet/txdetail/J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ) with code:
``` lisp
(coin.transfer-crosschain "k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab" "k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab" (read-keyset "receiver-guard") "0" 897.94089)
```

#### On master
```
$ time curl -v "$CWDURL/txs/search?maxheight=2991665&search=897.94089" | jq '.[] | { requestKey, height, code, pactId: .continuation.pactId }'
{
  "requestKey": "J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ",
  "height": 2991399,
  "code": "(coin.transfer-crosschain \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" (read-keyset \"receiver-guard\") \"0\" 897.94089)",
  "pactId": "J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ"
}

real	0m0,221s
```

`6nN3D_yO8uwDuoyRnD44eeEGOz8W5UCBVr2GYRrVSKI` is nowhere in sight.

#### On this branch

```
$ time curl -v "$CWDURL/txs/search?maxheight=2991665&search=897.94089" | jq '.[] | { requestKey, height, code, pactId: .continuation.pactId }'
{
  "requestKey": "6nN3D_yO8uwDuoyRnD44eeEGOz8W5UCBVr2GYRrVSKI",
  "height": 2991657,
  "code": "(coin.transfer-crosschain \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" (read-keyset \"receiver-guard\") \"0\" 897.94089)",
  "pactId": "J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ"
}
{
  "requestKey": "wjpfcLOT75MKSv2HyVYqbOvoKPN5Jt9OYKKxaBdiSyE",
  "height": 2991405,
  "code": "(coin.transfer-crosschain \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" (read-keyset \"receiver-guard\") \"0\" 897.94089)",
  "pactId": null
}
{
  "requestKey": "J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ",
  "height": 2991399,
  "code": "(coin.transfer-crosschain \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" \"k:49cea5ac09a05fb58f0ba609d6b746595d627064125958291973b4d14419aeab\" (read-keyset \"receiver-guard\") \"0\" 897.94089)",
  "pactId": "J-N2mC4lMffEbbphD9R0gjfDmKBu1GlHsEQr5pWC2nQ"
}

real	0m0,269s
```